requirements.txt
mcp
ibind
python-dotenv
uvicorn
starlette
httpx
pycryptodome
